TileMap Created
The tile map has been created with one small issues. The last texture to be created is the texture that will always be rendered no matter what for some reason. It does not seem to be saving a texture to the pointer but rather overwriting the pointers texture and just rendering the last texture to be assigned